added role setting option in emp page
